--- spark/src/test/scala/geotrellis/spark/io/http/HttpSlippyTileReaderTest.scala
@@ -1,28 +1,0 @@
-package geotrellis.spark.io.http
-
-import org.scalatest._
-import geotrellis.spark.SpatialKey
-
-class HttpSlippyTileReaderTest extends FunSuite {
-  	test("should return correct urls for given zoom level") {
-        val reader = new HttpSlippyTileReader[String]("http://tile.openstreetmap.us/vectiles-highroad/{z}/{x}/{y}.mvt")(fromBytes)
-        assert(reader.getURLs(2) == Seq("http://tile.openstreetmap.us/vectiles-highroad/2/0/0.mvt", 
-        								"http://tile.openstreetmap.us/vectiles-highroad/2/0/1.mvt",
-        								"http://tile.openstreetmap.us/vectiles-highroad/2/0/2.mvt",
-        								"http://tile.openstreetmap.us/vectiles-highroad/2/0/3.mvt",
-        								"http://tile.openstreetmap.us/vectiles-highroad/2/1/0.mvt",
-        								"http://tile.openstreetmap.us/vectiles-highroad/2/1/1.mvt",
-        								"http://tile.openstreetmap.us/vectiles-highroad/2/1/2.mvt",
-        								"http://tile.openstreetmap.us/vectiles-highroad/2/1/3.mvt",
-        								"http://tile.openstreetmap.us/vectiles-highroad/2/2/0.mvt",
-        								"http://tile.openstreetmap.us/vectiles-highroad/2/2/1.mvt",
-        								"http://tile.openstreetmap.us/vectiles-highroad/2/2/2.mvt",
-        								"http://tile.openstreetmap.us/vectiles-highroad/2/2/3.mvt",
-        								"http://tile.openstreetmap.us/vectiles-highroad/2/3/0.mvt",
-        								"http://tile.openstreetmap.us/vectiles-highroad/2/3/1.mvt",
-        								"http://tile.openstreetmap.us/vectiles-highroad/2/3/2.mvt",
-        								"http://tile.openstreetmap.us/vectiles-highroad/2/3/3.mvt"))
-    }
-
-    private def fromBytes(key: SpatialKey, arr: Array[Byte]) = ???
-}
